Move supported locales to utils
