+++ pages/login/index.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import 'tailwindcss/tailwind.css';
 import { SupportedLocale } from '../../@types';
 import { LoginTemplate } from '../../components/templates/LoginTemplate';
-import { staticEnglishProps } from '../../supportedLocales';
+import { staticEnglishProps } from '../../utils/supportedLocales';
 
 const LoginPage: NextPage<{ locale: SupportedLocale }> = (props) => (<LoginTemplate { ...props } />);
